requirements.txt
discord.py>=1.6.0
python-dotenv
pubchempy
rdkit
googletrans==4.0.0-rc1